+++ tests/helpers.py
@@ -6,7 +6,7 @@ class SettingsMock(ReloadableSettings):
     def __init__(self):
         self.checks = []
         self.notifiers = {}
-        self.creds = {}
+        self.creds = {'pass': 'password'}
 
     @classmethod
     def instance(cls):