+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+## Added
+- Browser form filling shorthand
 
 ## [3.0.10] - 2017-04-29
 ## Added